third_party/WebKit/LayoutTests/fast/css/lazy-parsing-first-letter-crash.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='stylesheet' href='data:text/css, p::first-letter {}'>
<p>Don't crash
<script>
test(() => { assert_true(true, "Should not crash"); });
</script>
